langchain-na-pratica/exercicios/capitulo_01/exercicio_2/pyproject.toml
[project]
name = "exercicio-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.8",
    "python-dotenv>=1.1.1",
]
